move to deduct
