+++ src/aioway/instrs/deducts/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) AIoWay Authors - All Rights Reserved
+
+from .deducts import *
+from .tensors import *
