+++ src/aioway/tspecs/__init__.py
@@ -1,6 +1,5 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 from .gym import *
-from .infers import *
 from .media import *
 from .tspecs import *
